Move PDF generation to about-to-submit callback
