+++ ccd-definition/CaseEvent.json
@@ -252,25 +252,13 @@
     "SecurityClassification": "Public",
     "CallBackURLAboutToStartEvent": "${CCD_DEF_CASE_SERVICE_BASE_URL}/callback/case-submission/about-to-start",
     "RetriesTimeoutAboutToStartEvent": "1,2,3,4,5",
+    "CallBackURLAboutToSubmitEvent": "${CCD_DEF_CASE_SERVICE_BASE_URL}/callback/case-submission/about-to-submit",
+    "RetriesTimeoutURLAboutToSubmitEvent": "1,2,3,4,5",
     "CallBackURLSubmittedEvent": "${CCD_DEF_CASE_SERVICE_BASE_URL}/callback/case-submission/submitted",
     "RetriesTimeoutURLSubmittedEvent": "1,2,3,4,5",
     "ShowSummary": "N",
     "ShowEventNotes": "N",
     "EndButtonLabel": "Submit"
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "ID": "attachSubmittedFormPDF",
-    "Name": "Attach submitted form PDF",
-    "Description": "Attach submitted form PDF",
-    "DisplayOrder": 1,
-    "PreConditionState(s)": "Submitted",
-    "PostConditionState": "Submitted",
-    "SecurityClassification": "Public",
-    "ShowSummary": "N",
-    "ShowEventNotes": "N",
-    "EndButtonLabel": "Save and continue"
   },
   {
     "LiveFrom": "01/01/2017",
