+++ ccd-definition/CaseEventToFields.json
@@ -312,17 +312,6 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseEventID": "attachSubmittedFormPDF",
-    "CaseFieldID": "submittedForm",
-    "PageFieldDisplayOrder": 1,
-    "DisplayContext": "OPTIONAL",
-    "PageID": 1,
-    "PageDisplayOrder": 1,
-    "PageColumnNumber": 1
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
     "CaseEventID": "addFamilyManCaseNumber",
     "CaseFieldID": "familyManCaseNumber",
     "PageFieldDisplayOrder": 1,
