+++ service/src/integrationTest/java/uk/gov/hmcts/reform/fpl/controllers/CaseSubmissionControllerAboutToSubmitTest.java
@@ -1,0 +1,98 @@
+package uk.gov.hmcts.reform.fpl.controllers;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.collect.ImmutableMap;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.OverrideAutoConfiguration;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.http.MediaType;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
+import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
+import uk.gov.hmcts.reform.document.domain.Document;
+import uk.gov.hmcts.reform.fpl.service.DocumentGeneratorService;
+import uk.gov.hmcts.reform.fpl.service.UploadDocumentService;
+import uk.gov.hmcts.reform.fpl.service.UserDetailsService;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.BDDMockito.given;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static uk.gov.hmcts.reform.fpl.utils.DocumentManagementStoreLoader.document;
+import static uk.gov.hmcts.reform.fpl.utils.ResourceReader.readBytes;
+
+@ActiveProfiles("integration-test")
+@WebMvcTest(CaseSubmissionController.class)
+@OverrideAutoConfiguration(enabled = true)
+class CaseSubmissionControllerAboutToSubmitTest {
+
+    private static final String AUTH_TOKEN = "Bearer token";
+    private static final String USER_ID = "1";
+    private static final ObjectMapper MAPPER = new ObjectMapper();
+
+    @MockBean
+    private UserDetailsService userDetailsService;
+    @MockBean
+    private DocumentGeneratorService documentGeneratorService;
+    @MockBean
+    private UploadDocumentService uploadDocumentService;
+    @Autowired
+    private MockMvc mockMvc;
+
+    @Test
+    void shouldReturnUnsuccessfulResponseWithNoData() throws Exception {
+        mockMvc
+            .perform(post("/callback/case-submission/about-to-submit")
+                .header("authorization", AUTH_TOKEN)
+                .header("user-id", USER_ID))
+            .andExpect(status().is4xxClientError());
+    }
+
+    @Test
+    void shouldReturnUnsuccessfulResponseWithMalformedData() throws Exception {
+        mockMvc
+            .perform(post("/callback/case-submission/about-to-submit")
+                .header("authorization", AUTH_TOKEN)
+                .header("user-id", USER_ID)
+                .contentType(MediaType.APPLICATION_JSON)
+                .content("malformed json"))
+            .andExpect(status().is4xxClientError());
+    }
+
+    @Test
+    void shouldReturnSuccessfulResponseWithValidCaseData() throws Exception {
+        byte[] pdf = {1, 2, 3, 4, 5};
+        Document document = document();
+
+        given(userDetailsService.getUserName(AUTH_TOKEN))
+            .willReturn("Emma Taylor");
+        given(documentGeneratorService.generateSubmittedFormPDF(any(), any()))
+            .willReturn(pdf);
+        given(uploadDocumentService.uploadPDF(USER_ID, AUTH_TOKEN, pdf, "2313.pdf"))
+            .willReturn(document);
+
+        MvcResult response = mockMvc
+            .perform(post("/callback/case-submission/about-to-submit")
+                .header("authorization", AUTH_TOKEN)
+                .header("user-id", USER_ID)
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(readBytes("fixtures/case.json")))
+            .andExpect(status().isOk())
+            .andReturn();
+
+        AboutToStartOrSubmitCallbackResponse callbackResponse = MAPPER.readValue(response.getResponse()
+            .getContentAsByteArray(), AboutToStartOrSubmitCallbackResponse.class);
+
+        assertThat(callbackResponse.getData())
+            .containsEntry("caseLocalAuthority", "example")
+            .containsEntry("submittedForm", ImmutableMap.<String, String>builder()
+                .put("document_url", document.links.self.href)
+                .put("document_binary_url", document.links.binary.href)
+                .put("document_filename", document.originalDocumentName)
+                .build());
+    }
+}
