+++ service/src/main/resources/HTMLTemplate.html
@@ -2198,7 +2198,7 @@
       </dd>
     </div>
   </dl>
-  <p class="consent">I, {{ case_data.userFullName }}, believe that the facts stated in this application are true.</p>
+  <p class="consent">I, {{ userFullName }}, believe that the facts stated in this application are true.</p>
 </main>
 </body>
 </html>
